[Bug][Zeta] SeaTunnelClient can not exit with error
